base_service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
python-dotenv==1.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing-extensions==4.10.0
uvicorn==0.27.1
websockets==12.0